third_party/blink/web_tests/editing/spelling/cold_mode_no_leak_detached_node.html

<!doctype html>
<title>Spellchecker should not keep an unbounded number of detached nodes alive</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<script>
// Spins spellchecker until idle. Return whether it's actually run.
function spinSpellChecker() {
  let spellCheckerIsRun = false;
  while (internals.idleTimeSpellCheckerState(document) !== 'Inactive') {
    internals.runIdleTimeSpellChecker(document);
    spellCheckerIsRun = true;
  }
  return spellCheckerIsRun;
}

test(() => {
  const n = 100;
  const numLiveNodesBefore = internals.numberOfLiveNodes();

  for (let i = 0; i < n; ++i) {
    {
      let input = document.createElement('input');
      document.body.appendChild(input);
      input.focus();
      assert_true(spinSpellChecker(), 'spellchecker is run');
      input.remove();
    }
    gc();
  }

  const numLiveNodesAfter = internals.numberOfLiveNodes();

  // The exact number of retained nodes might vary in runs. Give it a generous leeway.
  assert_less_than_equal(numLiveNodesAfter, numLiveNodesBefore + 40);
});
</script>
</body>
